SISRP-31764 Add search by CS ID to CalNet LDAP code
