+++ app/models/calnet_ldap/client.rb
@@ -71,6 +71,15 @@ module CalnetLdap
       results.first
     end
 
+    def search_by_cs_id(cs_id)
+      filter = cs_ids_filter([cs_id])
+      results = search(base: PEOPLE_DN, filter: filter)
+      results = search(base: GUEST_DN, filter: filter) if results.empty?
+      results = search(base: ADVCON_DN, filter: filter) if results.empty?
+      results = search(base: EXPIRED_DN, filter: filter) if results.empty?
+      results.first
+    end
+
     def search_by_uids(uids)
       [].tap do |results|
         uids.each_slice(BATCH_QUERY_MAXIMUM).map do |uid_slice|
@@ -91,6 +100,10 @@ module CalnetLdap
       uids.map { |uid| Net::LDAP::Filter.eq('uid', uid.to_s) }.inject :|
     end
 
+    def cs_ids_filter(ids)
+      ids.map { |id| Net::LDAP::Filter.eq('berkeleyeducsid', id.to_s) }.inject :|
+    end
+
     def search(args = {})
       ActiveSupport::Notifications.instrument('proxy', {class: self.class, search: args}) do
         response = @ldap.search args
@@ -103,7 +116,6 @@ module CalnetLdap
         end
       end
     end
-
 
     def tokenize_for_search_by_name(phrase)
       return [] if phrase.blank?
